refactor: POJO 객체를 도메인별로 패키지로 분리
