+++ src/main/java/com/hhpl/concertreserve/application/model/concert/ReservationInfo.java
@@ -1,4 +1,4 @@
-package com.hhpl.concertreserve.application.model;
+package com.hhpl.concertreserve.application.model.concert;
 
 import com.hhpl.concertreserve.domain.concert.type.ReservationStatus;
