+++ src/main/java/com/hhpl/concertreserve/application/model/concert/ScheduleInfo.java
@@ -1,4 +1,4 @@
-package com.hhpl.concertreserve.application.model;
+package com.hhpl.concertreserve.application.model.concert;
 
 
 import java.time.LocalDateTime;
