+++ src/main/java/com/hhpl/concertreserve/application/model/user/PointInfo.java
@@ -1,4 +1,4 @@
-package com.hhpl.concertreserve.application.model;
+package com.hhpl.concertreserve.application.model.user;
 
 import java.time.LocalDateTime;
