+++ src/main/java/com/hhpl/concertreserve/application/model/waitingqueue/WaitingQueueInfo.java
@@ -1,4 +1,4 @@
-package com.hhpl.concertreserve.application.model;
+package com.hhpl.concertreserve.application.model.waitingqueue;
 
 import com.hhpl.concertreserve.domain.waitingqueue.type.WaitingQueueStatus;
